+++ docs/source/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 v0.2.0
 ------
 
+* Added support for |prop_Msgpack| properties.
 * Added support for Json serialization of datetimes, Models and Keys.
 
 v0.1.0